requirement.txt
aioredis==1.2.0
asgiref==3.2.2
asn1crypto==0.24.0
astroid==2.3.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
autopep8==1.4.4
certifi==2019.9.11
cffi==1.12.3
channels==2.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.5
django-cors-headers==3.1.1
django-silk==3.0.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
enum-compat==0.0.2
eventlet==0.21.0
gprof2dot==2017.9.19
greenlet==0.4.12
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
mysqlclient==1.4.4
phonenumbers==8.10.18
Pillow==6.1.0
pycodestyle==2.5.0
pycountry==19.8.18
pycparser==2.19
pyfcm==1.4.7
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.4.3
PySocks==1.7.0
python-dateutil==2.8.1
python-engineio==3.10.0
python-socketio==4.3.1
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
twilio==6.30.0
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
zope.interface==4.6.0
